Added add edit stations fragment

+++ app/src/main/java/com/android/testtask/db/entity/Stations.kt
@@ -11,6 +11,6 @@ data class Stations(
     @ColumnInfo(name = "supplier") var supplier: String = "",
     @ColumnInfo(name = "fuelType") var fuelType: FuelType = FuelType.PETROL,
     @ColumnInfo(name = "qty") var qty: Int = 0,
-    @ColumnInfo(name = "sum") var sum: Long = 0,
+    @ColumnInfo(name = "sum") var sum: Double = 0.0,
     @ColumnInfo(name = "synced") var synced: Boolean = false
 )
